seating addition

+++ src/pages/movies.jsx
@@ -1,84 +1,72 @@
-import React, { useEffect, useState } from "react";
-import Layout from "../components/Layout/Layout";
+import React, { useEffect, useState, useContext } from "react";
 import { MovieProvider } from "../context/movieContext";
 import MovieForm from "../components/Movie/MovieForm";
 import fetchApi from "../utils/fetchApi";
 import MovieList from "../components/MovieList/MovieList";
+import { GlobalContext } from "../context/globalContext";
 
 function MoviesPage(props) {
-  const [role, setRole] = useState("");
-
-  let [message, setMessage] = useState("");
-  let [movies, setMovies] = useState([]);
+  const { user, movies, getMovie, message, setMessage } = useContext(GlobalContext);
   const [showEdit, setShowEdit] = useState(false);
   const [id, setId] = useState("");
-
-  useEffect(() => {
-    const ss = sessionStorage.getItem("role");
-    setRole(ss);
-  }, []);
 
   //getting all movies from backend
   useEffect(() => {
     getMovie();
   }, []);
 
-  async function getMovie() {
-    let res = await fetchApi.get("/movies");
-    //console.log(res.data);
-    if (res.data.statusCode === 200) {
-      setMovies(res.data.movies);
-      //setMessage(res.data.message);
-    } else {
-      console.log(res.data);
-    }
-  }
-
   //deleting a movie with ID
   const deleteMovie = async (id) => {
+    let movie = movies?.filter((m)=>{
+      return String(m._id)===id;
+    })
+   let con = window.confirm(`Are you sure to delete the movie ${movie[0].name}?`);
+   if(con){
     let res = await fetchApi.delete(`/movies/${id}`);
     //console.log(res.data);
     if (res.data.statusCode === 200) {
       getMovie();
-      //setMessage(res.data.message)
+      setMessage('');
     } else {
       console.log(res.data);
+      setMessage(res.data.message);
     }
+   }
   };
 
   //editing a movie with ID
   const editMovie = async (value) => {
-    //console.log(value);
     let res = await fetchApi.put(`/movies/${value._id}`, { ...value });
     //console.log(res.data);
     if (res.data.statusCode === 200) {
       getMovie();
-      setShowEdit(false);
       setId("");
-      //setMessage(res.data.message)
+      setMessage('');
+      setShowEdit(false);
     } else {
       console.log(res.data);
+      setMessage(res.data.message)
     }
   };
 
   //adding a new movie
   const addMovie = async (value) => {
+    // console.log(value);
     let res = await fetchApi.post("/movies/create-film", { ...value });
     //console.log(res.data);
     if (res.data.statusCode === 200) {
       getMovie();
+      setId('');
+      setMessage('');
       setShowEdit(false);
-      //setMessage(res.data.message)
     } else {
       console.log(res.data);
+      setMessage(res.data.message);
     }
   };
 
   return (
     <>
-      <div className="container-fluid">
-        <Layout />
-      </div>
       <div className="container">
         <MovieProvider
           value={{
@@ -90,39 +78,42 @@ function MoviesPage(props) {
             setId,
           }}
         >
-          <div>
-            {role === "admin" ? (
-              <div className="row">
-                <button
-                  className="btn btn-primary mt-4 ml-4"
-                  onClick={() => {
-                    setShowEdit(true);
-                  }}
-                >
-                  Add Movie
-                </button>
-              </div>
+          <>
+            {user.role === "admin" ? (
+              <>
+                <div className="row">
+                  <button
+                    className="btn btn-outline-primary mt-4 ml-4"
+                    onClick={() => {
+                      setShowEdit(true);
+                    }}
+                  >
+                    Add Movie
+                  </button>
+                </div>
+                <div className="row" style={{marginRight:"10%",marginLeft:"10%"}}>
+                  {showEdit ? (
+                    id ? (
+                      <div>
+                        <MovieForm message={message} id={id}></MovieForm>
+                      </div>
+                    ) : (
+                      <div>
+                        <MovieForm message={message}></MovieForm>
+                      </div>
+                    )
+                  ) : null}
+                </div>
+                {message ? <div className="text-danger">{message}</div> : ""}
+              </>
             ) : (
               ""
             )}
-            <div className="row">
-              {showEdit ? (
-                id ? (
-                  <div>
-                    <MovieForm id={id}></MovieForm>
-                  </div>
-                ) : (
-                  <div>
-                    <MovieForm></MovieForm>
-                  </div>
-                )
-              ) : null}
-            </div>
-            {message ? <div className="text-danger">{message}</div> : ""}
-            <div>
+
+            <div className="row m-0">
               <MovieList movies={movies} />
             </div>
-          </div>
+          </>
         </MovieProvider>
       </div>
     </>
